Replaced FFT routine by a version based on Temperton's GPFA
Calls to the 'four1' routine have been replaced by calls to
'fft_gpfa_ez' with the same arguments. This routine lives
in a new module in 'm_fft_gpfa.F', which also contains the
legacy code by Temperton and the original entry points.

The copies of radfft and bessph formerly inside the m_filter
module have been removed. This implies changes in the Makefile
for Genbasis.

Added a new script to clean the Util directory of object files
and executables.

added:
  Src/m_fft_gpfa.F
  Util/clean_all.sh
modified:
  Src/Makefile
  Src/m_filter.f90
  Src/radfft.f
  Src/recipes.f
  Util/Gen-basis/Makefile
  Util/Macroave/Src/Makefile
  Util/Macroave/Src/macroave.f
  Util/Macroave/Src/recipes.f