The pulay has been tested and it works ok!
dDscf converges well for the BN test!!

Added all the features needed to compute the scf-elements of the dynamical matrix.

Added dssmat and ddnaefs.

Everything works ok and I have reproduced the BN test!!

I have to Check times, subroutine callings  and clean the files and it will be done