dynamat.f: deleted some commented lines (previous versions)
cellxc.F90
lin 495 added re_alloc(mydDens) only if ddens is present.
lin 508 copymeshdata only if ddens is present.
lin 584 re_alloc dDleft/right elements if ddens is present.
lin 634 replicate the loop for the neighbor ddens

subroutine getGradDens, added new optional inputs to compute the gradient of the density or the perturbed density. Also it was modified the callings to this subroutine inside cellxc.