Added OMP in delrhok, updated printouts in linresscf, added Haux usage in delrhok
Added OpenMP to delrhok, for the test it is perhaps too small to really see a
difference. With only 72 orbitals it may be too small considering the overhead.
To do this I had to add preprocessor flags (renamed to .F).

Updated print-outs in linresscf to look more like the other siesta print-outs.

Enabled Haux usage (from densematrix) in delrho. However, currently this
is only used for the k-point variant as that is simpler.

Currently the OpenMP version cannot run if the number of threads exceeds the
number of "local" orbitals. However, for know this shouldn't be a problem
unless running KNL.

Removed k-point vs. gamma calculation in linres, this should be apparent
from the initial siesta print-out.