Updated delrhok with reduction of algorithm
I have also removed a line when solving the sub-space Hamiltonian
problem. It seems that eigenvalue problem solved was:
   H \psi = E (I + i I) \psi
but in the Gamma-only calculation it was
   H \psi = E \psi
I.e. if non-Gamma and the Gamma-point was calculated then
they would not coincide. I have taken the liberty of
removing the imaginary part of the overlap matrix.
Other than that, everything is the same.

NOTE: I ran the calculation by using (I + i I) and (I), they
both yield the same result in this case, but perhaps it is
because of the low occupancy of the state? Perhaps in
other systems it has higher importance?