rhooda.F: deleted some comments.
dhscf.F: bug in the dynamat calling for nspin=2. The dRho density (rho-rhoatm) was not well defined for nspin=2. Now, it enters to dynamat.F as sum of spins and difference. It looks ok!!

delrhok.f: found an inconsistency in the definition of rhoper which makes a bug for nspi==2. Rhoper has dimensions of RHOPER(MAXNH,NSPIN,3)

linresscf.f: blank lines...

At this point...LDA AND NSPIN==2 WORKS FOR BN TEST!!!!!!!!!!!!!!!!!!!!!!

*******************LIBXC**************************************
NSPIN==2 (spin Polarized fdf-option and PBE XC-potential)

ggaxc.f: changed the coefficient calculation needed for the PBE term.

cellxc.F90: PBE calculation term in the Vaux potential.

There are slight differences between the unpolarized and polarized case but looks OK!!! 

tested for BN and Si unit cells