Added calculation of bulk-current (per spin)
This now completes the bulk-velocity calculation by allowing
the calculation of the current flowing.
The current calculation can be enabled via
  BulkBias.Current true
but note that it has some performance implications.

To enable easy determination of the periodic directions I have
added a cell_periodic in the siesta_geom module. It allows
one to determine which directions are periodic regardless of
whether we are doing a Gamma calculation or not.

Also amended documentation.