Avoid corruption of 'elsi_get_edm' context when nscf=1
If the scf cycle only takes one iteration, iscf will be still 1 at the
time of the EDM retrieval call in post_scf_work. The initialization
logic in the ELSI module (based on iscf) will re-create all the data
structures, spoiling the context for the retrieval of the EDM.
This has been fixed for now by a minor kludge, pending a full refactoring
to treat more robustly the case in which 'Get_EDM_Only' is set.