Pass 'number-of-eigenstates' to the ELSI interface
This is a change from previous behavior. Instead of being hardwired
to a number slightly above the one corresponding to simple electron-counting,
the number of eigenstates to be computed in ELSI (for those solvers that
care about it) is now taken from the Siesta standard "number-of-eigenstates".
Note that, by default, this is the size of the basis, so for good performance
it should always be set (explicitly, or with the trick that specifies a given
number of states above the simple counting level, with a "-" sign).