maint: updated ARCH_EXPERIMENTAL
Now options should be 1 to enable a feature

E.g. WITH_MPI=1

This is necessary since users may expect WITH_MPI=0
to disable MPI, however this was not the case.
As long as it was defined (with *anything*) it was
enabled.
Now it is a bit more strict but more intuitive from
what I find in other build-systems.