Register trunk-qmmm code as of July 2012
This is the QM/MM code by Carlos Sanz, modified for easier working with Siesta:

* The 'driver' program is in Util/QMMM
* Some minimal changes have been made to the main source directory (Src)

See the file Docs/QMMM.CHANGES for more details, although there is not a lot
of documentation.

This commit corresponds to the bzr log stamp:

July 12 2012  Alberto Garcia  siesta-trunk-411-qmmm-4

although it was prepared by hand, without using a direct bzr to git translation.