maint: updated manual, typos and changed HAVE_WANNIER90
Updated manual to also contain the build instructions and
the revised build instructions. The new build instructions
are more in line with the other libraries.
I have also removed the reference to WANNIER90_INCFLAGS
in the Makefile. I think they just clutter the Makefile
system and we will get tons of these flags for various libraries.
I think we should go away from having *any* library specific flags
in the Makefile.

Changed all HAVE_WANNIER90 to SIESTA__WANNIER90. This is to
follow the rest of the libraries.

Also changed Util/Wannier90/README to reflect the above.