Reproduce Wannier functions with old-style Ylm convention
Siesta uses the Condon-Shortley phase convention for the spherical
harmonics.  This means, for example, that Y1,-1 is proportional to
'-y' and not to 'y'.  In general, the convention introduces a factor
(-1)^m, so spherical harmonics with odd m have a sign change with
respect to the "other" convention. It should be noted that the
Condon-Shortley convention has some advantages (see Wikipedia).

Basis orbital in Siesta inherit this phase convention. (So the positive lobe of
a 2py orbital is pointing towards -y)

Wannier90 uses implicitly the convention that the l=1, m=-1 orbital
should be proportional to 'y', etc. When Siesta orbitals are directly
used as "seed projection functions", this causes symmetry problems.

Rather than changing the low-level spherical_harmonics module which
encodes the phase convention, it is less intrusive to create new
projection functions as needed. For example, if one wants a Siesta
orbital with "py" character as projection seed, a new "twin orbital"
is created only for the purposes of Wannier90 operation. This twin
orbital is identical to the original, except that the radial part is
multiplied by -1. Thus the symmetry properties are restored.

The mechanics are in amn.F90. Twin "orbitals" are created on the fly
for odd m and registered as wannier projectors. Note that orbitals
with even m are also re-registered as wannier projectors (this is to
future-proof the scheme: modern versions of Siesta setup the matel
tables in a specific initial stage, instead of on-the-fly).

The resulting Wannier functions are completely identical as the ones
computed by the "non-Condon-Shortley" version, but, of course, their
coefficients on the orbitals with odd m have their signs changed,
since the basis orbital themselves differ by a sign. This can be checked with
the 'Util/readwfx" program, with a command like:

   readwfx -b1 -B1 -t 0.05  SrTiO3.manifold.1.WANNX

which means "get me the coefficients for band 2 of the WANNX file
whose modulus is above 0.05".

The .hr files are also identical (although zeroes sometimes appear
with a sign change).  The .tb files are identical for those elements
that are not very small (over 1.0e-14 or so).  This probably means
that there is precision loss somewhere along the way, unrelated to
this change.

Some extensions to the 'radial' and 'matel_registry' modules have been
made to implement this change.

The 'spher_harm' module has been reverted to its state in the master branch.