enh: Enabled Psolver library for 1.9.0 version
Keywords to control the cavity set-up in the soft-sphere model.
Keywords to manipulate gammaS, alphaS and betaV cavity parameters

Currently stress calculations are not functional. I am not exactly sure why.

The pre-processor flag is -DSIESTA__PSOLVER and we
force that Siesta is compiled with DP grid precision,
only if PSOLVER is requested compiled in.

I have also fixed the code to work with GridCell sampling.
Since atomic coordinates are used for cavity calculations
we have to use the correct atomic coordinates.
Now the atomic coordinates passed to dhscf are passed
down to poisson_psolver_init and poisson_psolver.

When calculating cavity epsilon for atoms we skip ghost-atoms.