Fix bug when mixing Siesta PAOs and Wannier trial orbitals
The logic for bookeeping of the placement of the projectors in cases such
as (in the WannierManifolds block):

        23 24 -1 26 -2

was faulty, and has now been fixed.

Note: The name 'trial_orbs' is used to refer to the "classic" Wannier
projection trial orbitals (as opposed to Siesta PAOs). This might
not be completely appropriate.