Jzero-local contribution in correspondence with QE
For Jharree & Jzero the difference on the `forward` direction for FFT
between the codes matters, leading to the adjustments in computations.

For Hartree flux: it flips the sign of the resulting elementary
product being integrated: `dimag(conjg(tc_dot(igp))*tc_v(igp)) * ...`

For Zero-local flux: the direction of the FT for `u_g` is reversed.