Add packed representation of ps generation valence configuration
For cosmetic reasons, it is convenient to have a way to report the
correct valence configuration used for pseudopotential generation,
notably when semicore states are present.

A new field 'gen_config_string' has been added to the Froyen
derived type in the ncps library. Its format is, for example

        5s:5p:6s:

This information is generated when translating PSML files, and
written to psf files if present. The Froyen reader has been
enhanced to read the information if present.