Enhanced pseudopotential information
* Better information for the valence charge for ps generation

The header information in the Froyen format (psf) records information
about the valence configuration for pseudopotential
generation. Froyen-derived programs (such as ATOM) cannot pseudize
multiple shells for the same angular momentum.  When semicore states
are present, ATOM uses an ionic configuration, with a reduced valence
charge. The legacy header, having one field per 'l', is still
appropriate. Other programs (such as ONCVPSP) can deal with multiple
shells per 'l', and the Froyen header cannot contain the full
information.

Currently, Siesta uses an intermediate data structure based on the
Froyen format when processing PSML files.  The existence and
identification of semicore states can still be done properly, since
they depend on the first pseudized shell per 'l'.  However, the total
valence charge at the time of pseudopotential generation has to be
handled in a special way, beyond adding up the contributions of the
different sections of the Froyen header. For PSML files this is
straightforward, using an extra component in the data structure.

PSML files from ONCVPSP can be 'converted' to the psf format for use
in previous versions of Siesta. This can be achieved by the 'psml2psf'
converter in the 'Pseudo/converters' directory. The converter adds
some extra data to the psf file (in a backward-compatible way) to
cover needed information, such as libxc codes. But up to now it had
not added information on the total valence charge for pseudopotential
generation, even though the proper field had been present already
('gen_zval', for the needs of the VCA tools).

Moreover, due to a bug, Siesta was not processing properly that extra
'gen_zval' field, even if set.

The 'gen_zval' value can be significant. Its difference from the
effective valence of the pseudo-ion is used by default as an 'extra
charge' for PAO generation, and it also determines the charge used
for KB projector generation.

The net result of the above is that, when a PSML file with semicore
states, but with a non-ionic configuration, was converted to the PSF
format, Siesta incorrectly set the 'gen_zval' value, forcing the use
of a non-zero effective charge for PAO generation. This led typically
to shorter orbitals.

The solution involves two changes:

- The psml2psf converter now records properly the total valence charge
  for pseudopotential generation in the appropriate field of the psf
  file. The program has been made more robust in the identification of
  the proper shells.

- Siesta now processes correctly the 'gen_zval' field (if present).

* More appropriate presentation of the valence configuration for
  pseudopotential generation, and of pseudized shells

Further cosmetic changes are needed to avoid displaying incorrectly
the valence configuration for pseudopotential generation when programs
other than ATOM produce the pseudopotential, notably when semicore
states are present. It is appropriate to clearly differentiate the
list of 'pseudized shells' from the 'valence configuration' in the
report.

A new field 'gen_config_string' has been added to the Froyen derived
type in the ncps library. Its format is, for example

            5s:5p:6s:

This information is generated when translating PSML files, and written
to psf files if present. The psf reader has been enhanced to read
the information if present.

* Better presentation of the information regarding semicore states

It is now more compact. Also, information is now given for the
converse case, that of a shell that should nominally be 'valence', but
that is kept in the 'core' manifold. This could be the case of Cu, for
which the 3d shell is considered to be 'valence', but which could
potentially be left in the core.

The semicore report is printed even if the PAO.Basis block is specified.

* Removed spurious output lines from reader and dumper subroutines.

* Re-organized the pseudopotential-related information in the output file
  so that each species' contribution is more clearly framed.