Store nnz_g in sparse_matrices module, for reuse in ELSI
Computation of nnz_g, in the ELSI setup phase, and elsewhere, needs
an 'all_reduce' operation.

The variable is now stored in the sparse_matrices module by 'state_init',
and re-used in the Gamma-point execution of elsi_getdm.

(The original rationale for this was a strange delay in MN4 runs for
large numbers of processors)