enh: changed input, fixed compilation and shipped lapack
Wannier90 input has drastically changed.

Now instead of single blocks containing indices for referencing
everything is now divided by names.
This makes it a bit clearer where stuff in coming in and what
each line has of meaning.

I have updated the single test (wannier90_in_siesta) to reflect
this. It runs.
TODO update manual to reflect this, I wait for Javier's comments.

I deleted the broadcast functionality. The old reading on IO-node,
then B-casting is obsolete and should never be done.
Now everything is read in parallel.
This cleaned up the code quite a bit since the reading and
post-processing of the read data could be done in one go.

Some of the additional stuff may still be leveraged for name-block
inputs, but so far this is just to show Javier what can be done.
Also, we need to ensure that defaults are handled. Currently,
I don't check for this.

I deleted the vectorproduct function, it was already implemented
in cross.f.

I changed the block kMeshForWannier to Wannier.k. It may now
also either be given as a list or a block.
Since the manifolds now have explicit names, so does the output.
This is probably a bit easier to comprehend since many manifolds
can be tricky to handle. This will also make it easier
to disable single manifolds (simply comment them out in the
WannierManifolds block).

A couple of files did not allow Siesta to compile without additional
definition flags, these are now added to allow compilation.

The wannier90 library uses some non-existing LAPACK routines in the
shipped LAPACK library. This is now fixed.