enh: cherry-picked Alberto suggestions, and checked cavity
Amended fix suggested by Alberto.
This fixes the stress tensor by correcting with 2U/V
for the diagonal components. We are not really sure
whether this affects the off-diagonals.
Also ensured that the potential is always positive.
We are also not sure what causes this.

Additionally tested a cavity calculation.
It seemed to require additional calls to the PSolver
machinery to extract the forces from the cavities.
Since the internal PSolver calculates the forces
including the electrostatic potential we have to
do the call twice, 1) calculate only electrostatic
forces, 2) change sign and calculate electrostatic
+ cavity forces.
This seems to yield reasonable forces, but I am
not sure at all that they are in fact correct.

Also enabled CUDA accelerators (but not tested).