Modifications by Carlos Sanz to support QMMM
These modifications (as of June 8, 2011) were done by Carlos Sanz
over the 'siesta-trunk-377' version to support the interface to the
QMMM driver.

Basically, the allow the reading and processing of a supplementary
external potential created by the 'classical' ions, and modify
the 'iopipes' interface to process also a 'relaxd' variable to
signal the end of the process.

There were also modifications done on the 3.X branch, at around the
same time.