Modifications to support QMMM (ca. July 2012; rebased from trunk-411 version)
These modifications are the latest by Carlos Sanz (as of July 2012)
based on 'trunk' (i.e., not 3.x)  to support the interface to the QMMM
driver. They are based on 'trunk-411' and rebased here.

The original set of changes, over 'trunk-377' (circa June 2011), were
intended to basically, allow the reading and processing of a
supplementary external potential created by the 'classical' ions, and
modify the 'iopipes' interface to process also a 'relaxd' variable to
signal the end of the process.

(A change which might need to be re-considered is to add a new
argument "iscf" to dhscf. This forces changes in a number of other
routines which are not directly related to any QMMM functionality.)

Carlos Sanz maintained also modifications done on the 3.X branch, at
around the same time. The basic difference is that here there is extra
logic to avoid reading the external potential repeatedly during the
scf cycle, but there are other subtle differences that need to be
checked, as the version that has been more thoroughly tested in
production is that based on the 3.X branch.