Modifications to support QMMM (circa July 2012; trunk-411 version)
This set of changes are the 'trunk' version (as of July 2012)
of the modifications needed to support the interface to the QMMM driver.

The original set of changes in this branch, over 'trunk-377' (circa
June 2011), were written by Carlos Sanz in parallel to his changes to
the 3.X version. (Two sets of changes were written because the 'trunk'
version was going to incorporate some relevant optimizations to the
grid operations.)

They are intended to, basically, allow the reading and processing of a
supplementary external potential created by the 'classical' ions,
modify the 'iopipes' interface to process also a 'relaxd' variable to
signal the end of the process, and write the total charge.

We have imported the changes directly from the final state of the
corresponding bzr branch (trunk-411--qmmm-4), as there were some
clarifications and enhancements over the initial set of changes by
Carlos Sanz. We have removed the QMMM driver program, which now lives
in a different repo.

(A change which might need to be re-considered is to add a new
argument "iscf" to dhscf. This forces changes in a number of other
routines which are not directly related to any QMMM functionality.)