mnt: clean-up in SCF loop and writing of HSX
The initial SCF step included the timing of the
initial setup_hamiltonian call.
One would want the IterSCF time-stamp for the first
SCF to be able to estimate the correct timing for
a number of SCF steps. This should be accomblished with this
PR.

In compute_dm I have moved some PEXSI-only calls inside the
isolve == SOLVE_PEXSI segment to clarify their belonging.
This also removes the Ebs call on every SCF loop
which really wasn't necessary for anything but PEXSI.

Writing the HSX files is now limited to only happen at the end of
the SCF cycle if converged. This makes more sense in terms of
COOP analysis etc. since a non-converged simulation can't retain
a good enough H.
The change is now such that:
- HSX is written if h_setup_only is true.
- HSX is written at SCFconverged if savehs OR write_coop

This also fixes the case that the correct H is written for
mixing H when savehs was false and write_coop was true.