bug: fixed writing LDOS grid to separate NetCDF file
Fixes #95. Problem was that NetCDF writing the grid
had hard-coded the variable name for storing the
data to RHO, but this same code was used for LDOS calculation.

This should now be fixed.
While the solution isn't elegant, I think we can use this until
we make a better solution for the entire grid saving mechanism.