bug: removed shifting EDM when reading old TSDE
This only affected forces on electrode atoms.
The Fermi-level shift was a means to figure out what
the potential offset was for the electrode compared to the
device. However, it does not work for all systems.
Hence we remove it.
The forces on electrodes is still *wrong* when the EDM
is read from the electrode TSDE file. However, the forces
on the electrodes were never to be trusted anyways and
should never be used (this is why TSFA is written!).

Increased precision write-outs a couple of places.

Added Hartree FFT fix value to the NetCDF output.