Further changes to the build directory documentation.
This commit incorporates the corrections and improvements
from MR !55 by Nick and Alberto.

It also provides further clarifications and combines information
that was available on Obj/README and the manual, so that both
sources of documentation essentially contain the same building
instructions.

Util/README is also updated to point the user to the instructions
available in Obj/README on how to compile utilities.