Switch to non-perturbative polarization scheme automatically
The program will now switch automatically to using a non-perturbative
scheme to generate polarization orbitals for a given species, if there
are lower-lying orbitals with the same l.

The perturbative scheme can still be enforced by putting the species
in the PAO.PolarizationScheme block. For example:

   %block PAO.PolarizationScheme
     Na perturbative
   %endblock PAO.PolarizationScheme

will ignore any program heuristics and use the perturbative scheme.