enh: streamlined memory access for delkmatgen and ensured nullification
A possible problem was discovered with teh delkmat matrices.
They were never nullified while they were tested for associtivity.
This may result in undefined behaviour. The best practice is to
define them with the null() assignment.

The data delkmatgen is now transposed meaning that memory access
will always be fast. Possibly this was why Intel compilers
had a hard time compiling and resulting in temporary arrays etc.

In principle the delkmat never needs to be allocated. It can
point directly to the corresponding delkmatgen index.