enh: added patch for v3.1.0 reran test example
The wannier90_in_siesta test has been runned with v3.0.0
and v3.1.0 using the patches in the Util/Wannier90 folder.
Both yield exactly the same output with no other warnings.

I think it is safe to say that they can both be used.

Note that I have merged the two patches.
Although one is for making it work in siesta and the other
a bug, I think it is vital that the patches are simple for
users. So combining to 1 is better.