bug: fixed decoupling degenerate states
Now the decoupling algorithm only takes into
account the bulk-bias direction. This will be the
important parameter.
This actually turns out to be much faster since
one need only setup the dk matrix once per
k-point if there are degenerate eigenvalues.
This is now handled via a routine global variable.

The test has been runned again and the agreement with
the conductance quantum is better.