Restore 'use sys' idioms to minimize changes wrt master
Some files had changed in the psml branch only because 'die',
'message', and 'bye' had been re-implemented as external routines, and
the 'sys' module dependency removed in most places.

Now, however, a spurious dependency on cml_subs in the sys module has
been removed, and a mechanism for the use of external routines fully
implemented.

This has removed changes in a few files.

In the Utils, some 'local_sys' files have been removed and the
dependencies of Siesta routines taken care of by the inclusion of
the top-level 'sys.o' in the objects lists. In these cases the
'handlers.f' file now contains copies of the (possibly dummy) 'die',
'message', and 'bye'.  In other cases, the dependencies are only
(typically) on 'die', and a 'local_sys' file has been retained.