mnt: sorted BTD matrix output
It is not necessary for end-users to see
the entire BTD format. Instead we only print
the blocks in a sorted fashion. This is of course
not *true* in the sense that internally the BTD is
very different. However, it may give users a hint with
a much reduced output for large matrices.