mnt: changed default eta value for TS/TBT
For very large structures the eta value for the device
region is quite important when doing spectral function
calculations. Now we use a non-zero default value for the
device region eta values for transiesta and tbtrans.

It is defaulted to minval(Elecs(:)%eta)/100 to always
ensure electrodes have the largest smearing unless
user-defined.

This is important for cases where localized states
exists and is a way around these problems.