mnt: cleaned transiesta tests and changed eta value for TS/TBT
The tests were really annoying to track and follow.
Instead of having names that do not follow the folder
names, we now adhere to the regular test requirements
that uses the folder name as the test name.
In the case of transiesta we also have electrode calculations.
So now all files has the folder name as prefixes.