Add gen-basis target
This is an example of the use of fdf and xc by a Util program.

Top-level file dependencies are listed explicitly.

The CMake work has uncovered a missing explicit dependency on
'cml' in the legacy Makefile through 'die'. It has been solved
by creating a new local file 'local_die.f'.

Note that this target can link to the 'mpi' version of SiestaXC
even without any explicit MPI definitions here (since 'atomxc' does
not need any MPI).

It links to the serial fdf.