Add netCDF support
The discovery of the netCDF libraries is done using the pkg-config
interface. This is not completely foolproof, and there should be
a fallback (maybe using environmental variables).

Siesta and TBtrans are now enabled for netCDF (and actually for NCDF
and NCDF_4) if netCDF is requested in the cmake line:

FC=XXXX   cmake .. -DWITH_NETCDF=ON --DCMAKE_GENERATOR=Ninja [...]