Fix (syntax): specify format of explorer outputs.
It is not required that values produced by list-directed output are
separated by blanks, they may be separated by a comma (or a semicolon
if the decimal edit mode is COMMA). The processor may also replace a
sequence of identical values by a repeated constant of the form r*c.
[See, e.g., subsection 10.9.2 List-directed output of the Fortran 2003
standard.]

This commit specifies the format of the outputs of the explorer binaries
when multiple values are to be written, in order to ensure that they
will be parsed correctly by the generate.sh script.

By doing this, this commit allows MPI-enabled compilations with the CRAY
Fortran compiler.