Fixed tensorial issues in LHS of linear equation
In the left-hand-side of the linear equation for the determination
of psi_hat_c, the operator:

     H - eps*1 + alpha_P_v

is represented properly in matrix form as:

  S_inv * H - eps*1 + alpha*(0.5*DM)*S

taking into account the non-orthogonality of the basis set.

With this (and the use of a general (not symmetric) linear solver),
the resulting psi_hat_c passes the test of being in the unoccupied
subspace.

Jele results are closer to the target, but about 5% off. This might be
due, among other things, to the fact that [H,r] is computed for the
current coordinates, and not for the "base step".