Inherit 'n' quantum numbers from basis_specs routine
The setting of cnfigtb in 'atom' was plagued by potential
inconsistencies. Now, a new array 'nprin' (set in 'basis_specs')
is passed to 'atom', and cnfigtb is set from it.

(This has clarified a number of issues. For example, that polarization
shells specified in PAO.Basis can appear before lower-lying
('semicore') shells in the 'nsm' list).