Sort shells by n quantum number when using PAO.Basis
When semicore shells were present and (explicit) higher-lying
shells with the same l appeared earlier in the PAO.Basis block,
the program would mistakenly interchange the shell specifications.

This is an example (for a Ge pseudo with 3d as semicore):

%block pao-basis
Ge 4
n=4  0  2
 7.4  0.0
 1.0  1.0
n=4  1  2
 8.56   0.0
 1.0    1.0
n=4  2  2 Q 1.5
 8.0    0.0
 1.0    1.0
n=3  2  1
 3.18
 1.0
%endblock pao-basis

A sorting step has now been implemented to fix this.

Note that polarization shells declared implicitly with the 'P'
idiom in the polarized shell did not cause problems.