enh: moved a few variables into the w90 type and cleaned
Updated the w90_in_siesta test.

Ensured each manifold can now control the output of the UNK
and the threshold individually. This was very easy and should
be easier for end users. The global controls still exists
under a different name.

Changed global wannier90 input option names. There is no need
to not prefix with Wannier as that may only confuse users.
Even if they are developer only, they should be concise and
meaningful. This should now be obeyed.

Documentation has been prepared for input from Javier.
Now all flags are basically described and all that is required
is a re-construction of the documentation to fit the new input
mechanism.