Work around pointer initialization problem
The new procedure-pointer code in alloc.F90 is for now wrapped with

  #ifdef WITH_ALLOC_PROC_POINTER

If F2008 cannot be used, the memory event procedure pointer is set to
null() at initialization (the only option available), and there is a
check in alloc_count for the association status.  If unassociated, it
is associated to the 'simple' version inside the alloc module, which
does nothing.

If F2008 is enabled  (-DF2008) the pointer is directly initialized
to the 'simple' routine.

In both cases, siesta_init calls a new setter routine to point the
memory handler to the appropriate Siesta routine.

If WITH_ALLOC_PROC_POINTER is not set, the old code still works, with
the external reference (new _long version, for i8b integers) in
file alloc_memory_event.F. Note that the old version of the handler
routine is still there, to resolve gridxc's handler. The library is
going to be updated to proc pointers soon, but older versions will
need to be supported.

-DWITH_ALLOC_POINTER and -DF2008 can be set in arch.make, using
the newly processed WITH_FPPFLAGS symbol.