mnt: moved some memory types into the alloc module
Also fixed a bug for COMPLEX(4) memory allocation.
Using the common specifiers 'C' and 'Z' for complex
and double complex I think should be embedded.
The complex(4) was using type = 'S' and factor 2
which meant a memory count of 2*size(array) bytes
which is a factor 4 wrong.

This fixes the complex(4) cases, and re-inserts
the 'C' and 'Z' as valid type specifiers for memory.

Note that the code has a big inconsistency in the
use of the 'S' type, see e.g. memory_log vs. alloc
for the 'S' type. 1 vs. 4 bytes per element.
It is not clear to me what the use-cases are.