mnt: beautified spin moment output
The old output was not easy on the eyes and it was difficult
to separate x, y, z from total. With this new change
it is clear which numbers belongs to a vector, and which
number is the magnitude. Due to the extra characters it will
be easy for scripts to accommodate both things.