Streamlined treatment of internal libraries
A new section in build.mk handles built-in libraries (currently: fdf,
ncps, psop, MatrixSwitch) and generates *_INCFLAGS and *_LIBS symbols
in a manner compatible with the usage for external libraries.

Each .a file will reside in the appropriate subdirectory of the
compilation directory (Obj by default).

The only difference with respect to external libraries is that client
codes (Siesta or any of the Util programs) need to include a
dependence (say, 'prog: $(FDF_LIBS)' to trigger the compilation.

Utilites can now be compiled in any order without having compiled
Siesta first.

Notes:

 - The MPI interface code could potentially be treated in the same way.
 - The FDICT and NCDF libraries could also be handled like this.