(wip) Use libfdf code, including 'client-units' functionality
Src/fdf is bypassed in favor of Src/libfdf, which should contain (either
through a symlink or a git submodule) the libfdf tree.

A version of libfdf with the 'client-units' functionality is needed.

The build.mk file takes care to process the fdf symbols adequately.

The 'units' file in Siesta now conforms to the latest CODATA numbers.