Make fdf MPI-agnostic. Streamline treatment of internal libraries
A client-side broadcast of the fdf data structure allows MPI programs
to use a new, MPI-agnostic, fdf library. There is thus no need to compile
specific MPI or serial versions of the library for each client program.

Utility programs can now avoid recompiling fdf for their own needs. A
single rule, and the appropriate symbol handling, in the build.mk file
further simplifies the use of the library. This scheme has been
extended to other "internal" libraries, such as ncps, psop, and
MatrixSwitch, and also to the MPI interfaces and the built-in linear
algebra library.  It could also be extended to fdict, and maybe to
ncdf, but the details have to be worked out.

The build.mk file is currently included at the end of the arch.make
file, and a checker snippet is included by every makefile which needs
to use the build.mk logic, to make sure that it is indeed
included. The same 'include' approach could perhaps be used to simply
incorporate 'build.mk' itself after arch.make, and avoid burdening the
users with the current cryptic lines.

Both build.mk and the checker snippet are put in the object directory
by the Src/obj_setup.sh script. Fine-tuning modifications by users are
possible but will be overwritten by a further execution of the
Src/obj_setup.sh script. A warning or a backup procedure can be
implemented if deeemed necessary.

One remaining source of confusion in the compilation of the utility
programs is the abuse of COMP_LIBS, and LIBS, which are in principle
needed by SIESTA. Perhaps every program should explicitly list the
libraries needed, with standard names. The same problem appears when
SIESTA "features" (i.e., WITH_MPI) have to be deactivated for a given
utility. Here the "features" are maximal features as set for SIESTA,
and the utilities inherit them. It should be possible now for a given
utility to be compiled with MPI, say, even though SIESTA itself is not.

Extensions and further ideas:

- Put the utilities under the same "object directory" as SIESTA, and
  compile and install them appropriately.

- Bring "external" libraries into the "internal" model through submodules
  subtrees, or subrepos, with appropriate entries in build.mk.