Change strategy for building Util/COOP targets
Ninja complains about "several paths" for a given .mod file if there are
common source files for several targets. (Not always: individual targets
build just fine).

The workaround is to create a pseudo library (OBJECT) with the common sources, but
it is ugly and will not work if at some point different targets might want to
define different pre-processor flags, for example.