Add Src/easy-fdict and Src/easy-ncdf
Create simple sources for the FDICT and NCDF targets, and streamline
the dependencies.

There is a single ncdf library. When NCDF_PARALLEL is in effect, in some cases
(e.g. Util/TS/tshs2tshs) the executable has to be linked with MPI even
if it does not use MPI itself.