Set explicitly CMAKE_INSTALL_LIBDIR to 'lib' (mk-based scheme)
Following the GNU conventions, CMake can set this string to either
'lib' or 'lib64', depending on the platform, and discovering the right
value outside CMake is not trivial. Since in the makefile-based scheme
we are not really installing the libraries, it is more expedient to
force the value to something concrete ('lib').