Make sure that subordinate CMake's use arch.make's compilers
(This affects only the {mk + deps on-the-fly} scheme)

By passing FC_SERIAL or FC_PARALLEL as CMAKE_Fortran_COMPILER to the
dependencies' CMake invocations we guarantee that the same compilers
are used. In particular FindMPI will use the value of the above
variable as a hint for the MPI location.