Allow specific flags (no-opt) for certain source files
By setting a property in some source files, they can be
compiled with flags that are different from the default for
the target. This can be used to compile certain files in Src
with no optimization (particularly for the Intel compiler).

The (added) flags are those in Fortran_FLAGS_DEBUG (options appearing
later in the command line take precedence). The DEBUG build type is
here abused to mean "no optimization". A new build type "Check" (and
its attendant flags) has been defined to mean "use extra checks"
(e.g. -fcheck=all).  This should be refined. Maybe a new build type
"NoOpt" could be defined as well.